+++ php/api/routes/progress.php
@@ -1,0 +1,5 @@
+<?php
+declare(strict_types=1);
+
+$user = Auth::requireUser();
+Http::send(200, ['modules' => Progress::modules($user['id'])]);